requirements.txt
Django>=5.0,<6.0
django-allauth>=65.0.0
django-htmx>=1.17.0
django-filter>=24.0
django-tailwind>=3.8.0
django-widget-tweaks>=1.5.0
WeasyPrint>=62.0
openpyxl>=3.1.0
psycopg2-binary>=2.9.0
gunicorn>=21.0.0
dj-database-url>=2.1.0
markdown>=3.5.0
